changes for pretext-plus
