+++ .changeset/brave-pugs-shave.md
@@ -1,0 +1,5 @@
+---
+"@pretextbook/import": minor
+---
+
+Export the paste placement helpers (`placeConvertedMarkup`, `isInlineContext`, `reindentForContext`, `firstDivisionTag`), moved here from the VS Code extension so every host fits converted markup to the cursor the same way.
