+++ packages/import/package.json
@@ -54,10 +54,10 @@
     "test": "vitest run"
   },
   "dependencies": {
-    "@pretextbook/format": "*",
-    "@pretextbook/latex-pretext": "*",
-    "@pretextbook/latex-style-pretext": "*",
-    "@pretextbook/remark-pretext": "*",
+    "@pretextbook/format": "^0.5.0",
+    "@pretextbook/latex-pretext": "^0.3.0",
+    "@pretextbook/latex-style-pretext": "^0.5.0",
+    "@pretextbook/remark-pretext": "^0.1.0",
     "@tailwindcss/vite": "^4.0.0",
     "jszip": "^3.10.1",
     "tailwindcss": "^4.0.0"
