+++ packages/latex-pretext/package.json
@@ -49,7 +49,7 @@
     "tslib": "^2.3.0"
   },
   "devDependencies": {
-    "@pretextbook/ptxast": "*",
+    "@pretextbook/ptxast": "^0.2.0",
     "xast-util-from-xml": "^4.0.0"
   },
   "publishConfig": {
