+++ packages/latex-style-pretext/package.json
@@ -39,7 +39,7 @@
     "vscode-languageserver-types": "^3.17.5"
   },
   "devDependencies": {
-    "@pretextbook/latex-pretext": "*",
+    "@pretextbook/latex-pretext": "^0.3.0",
     "@pretextbook/unified-latex-to-pretext": "^0.0.13",
     "@unified-latex/unified-latex": "^1.8.4"
   },
